Divided prog into sev. files; edited ImageIterator class; added exception handlers
